refactored and organized
