+++ cnwi/td.py
@@ -31,43 +31,3 @@ class TrainingData:
     
         labels = self.collection.aggregate_array(self.label).distinct().sort().getInfo()
         return {lbl: value for value, lbl in enumerate(labels, start=1)}
-
-
-def generate_samples(stack: ee.Image, training_data: TrainingData, scale: int = 10,
-                     projection = None, tile_scale: int = 16, geom: bool = False) -> Union[TrainingData, ee.FeatureCollection]:
-    
-    ee_lookup = ee.Dictionary(training_data.labels)
-    
-    def insert_value(element: ee.Feature):
-        return element.set(training_data.value, ee_lookup.get(element.get(training_data.label)))
-    
-    def add_geometry(element: ee.Feature):
-        x,y = element.get('POINT_X'), element.get('POINT_Y')
-        return ee.Feature(ee.Geometry.Point([x, y])).copyProperties(element)
-    
-    to_sample = training_data.collection if isinstance(training_data, TrainingData) else training_data
-    
-    samples = stack.sampleRegions(
-        collection = to_sample,
-        properties = [training_data.value, training_data.label, 'POINT_X', 'POINT_Y'],
-        scale = scale,
-        projection = projection,
-        tileScale = tile_scale,
-        geometries = geom
-    ).map(insert_value).map(add_geometry)
-    
-    training_data.samples = samples
-    
-    return training_data
-
-
-def label_frequency(gdf:gpd.GeoDataFrame, label: str) -> int:
-    summary = gdf.groupby(label)[label].count()
-    if len(set(summary)) > 1:
-        return summary.min()
-    else:
-        return None
-
-
-def balance_dataset(gdf, n: int, label: str, random_state = 1) -> gpd.GeoDataFrame:
-    return gdf.groupby(label).sample(n=n, random_state=random_state)
